requirements.txt
python-dotenv~=1.0.1
pandas==2.2.2
selenium==4.23.1 
beautifulsoup4==4.12.3
pymongo==4.8.0